make user verify before deleting functions

+++ python/scripts/delete_lambdas.py
@@ -7,5 +7,6 @@ REGION = "us-west-2"
 client = boto3.client("lambda", REGION)
 response = client.list_functions()
 for function_info in response["Functions"]:
-    client.delete_function(FunctionName=function_info["FunctionName"])
-    print("Deleted %s." % function_info["FunctionName"])
+    resp = raw_input("Delete %s?\ny/n: " % function_info["FunctionName"])
+    if resp == "y":
+        client.delete_function(FunctionName=function_info["FunctionName"])
